Move test stuff out of UI utility
